fix: problem with displaying the file popup
So it was because an error in browser console said "Mismatching childNodes vs. VNodes".
According to this issue thread: https://github.com/nuxt/nuxt.js/issues/1594#issuecomment-424805359
It was because v-if causes such problem, so I replaced all the v-if with v-show instead, and see how it goes.
By the way, this bug isn't a problem in development mode, but it is in production mode. You can click the link for more details.